requirements.txt
streamlit
google-generativeai
pandas
openpyxl
pillow
PyMuPDF
pdf2docx
selenium
webdriver_manager
requests
send2trash
python-docx
docx2pdf
pyinstaller
pytesseract
streamlit-option-menu
streamlit-aggrid
pdfplumber
plotly
watchdog
streamlit-elements
fastapi
uvicorn
python-multipart
